[feat] optimise

+++ GameManager.cpp
@@ -16,7 +16,7 @@ void GameManager::Load()
 
 void GameManager::Update()
 {
-	Vector2D mousePosition = { (int)GetMousePosition().x, (int)GetMousePosition().y};
+	Vector2D mousePosition = { GetMousePosition().x, GetMousePosition().y};
 
 	if (!mCanClick)
 	{
@@ -27,15 +27,14 @@ void GameManager::Update()
 		if (IsMouseButtonPressed(MOUSE_LEFT_BUTTON))
 		{
 			//std::cout << "START AT : " << mousePosition.x / ((int)GetScreenWidth() / MAP_SIZE) << " X | " << mousePosition.y / ((int)GetScreenHeight() / MAP_SIZE) << " Y" << std::endl;
-			mMapManager->SetStart({ mousePosition.x / ((int)GetScreenWidth() / MAP_SIZE), mousePosition.y / ((int)GetScreenHeight() / MAP_SIZE) });
+			mMapManager->SetStart({ mousePosition.x / (GetScreenWidth() / MAP_SIZE), mousePosition.y / (GetScreenHeight() / MAP_SIZE) });
 		}
 		else if (IsMouseButtonPressed(MOUSE_RIGHT_BUTTON))
 		{
-		//	std::cout << "END AT : " << mousePosition.x / ((int)GetScreenWidth() / MAP_SIZE) << " X | " << mousePosition.y / ((int)GetScreenHeight() / MAP_SIZE) << " Y" << std::endl;
-			mMapManager->SetEnd({ mousePosition.x / ((int)GetScreenWidth() / MAP_SIZE), mousePosition.y / ((int)GetScreenHeight() / MAP_SIZE) });
+		    //std::cout << "END AT : " << mousePosition.x / ((int)GetScreenWidth() / MAP_SIZE) << " X | " << mousePosition.y / ((int)GetScreenHeight() / MAP_SIZE) << " Y" << std::endl;
+			mMapManager->SetEnd({ mousePosition.x / (GetScreenWidth() / MAP_SIZE), mousePosition.y / (GetScreenHeight() / MAP_SIZE) });
 		}
 	}
-
 }
 
 void GameManager::Draw()
